feat(app): 기본 쿼리 파라미터 지정 훅을 리네이밍하고 useEffect 를 쓰지 않도록 개선
